pyproject.toml
[tool.poetry]
name = "anti-messenger"
version = "0.1.0"
description = "Picture messaging application."
authors = ["Manuilov Daniil <manuilovdaniil@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "anti_messenger"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-environ = "^0.10.0"
mysqlclient = "^2.2.0"
pillow = "^10.0.0"
django-tinymce = "^3.6.1"
django-registration = "^3.4"
pytesseract = "^0.3.10"
opencv-python-headless = "^4.8.0"
crispy-tailwind = "^0.5.0"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
py_version = 311
profile = "django"
known_first_party = ["users", "communications"]
skip = ["*/migrations/*"]

[tool.flake8]
max-complexity = 10
max-line-length = 120
exclude = "venv/,*/migrations/"
per-file-ignores = [
    "*/migrations/*: E501",
    "venv/*: E501"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "anti_messenger.settings"
python_files = "test_*.py"
